start a new test file for testing the cyl Operators
